pyproject.toml
[tool.poetry]
name = "iot-gateway"
version = "0.1.0"
description = "Raspberry Pi BLE/BLUNO -> SQLite + Arduino SIM7070G MQTT bridge"
authors = ["Tu Nombre <tu@correo>"]

# 👇 IMPORTANTE: packages va AQUÍ, no en scripts
packages = [{ include = "gateway", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pyserial = "^3.5"
aiosqlite = "^0.20.0"
pydantic = "^2.8.2"
PyYAML = "^6.0.2"
typer = { version = "^0.12.3", extras = ["all"] }
tenacity = "^8.5.0"
bluepy = "^1.3.0"
bleak = "^0.22.2"
click = ">=8.1.7,<8.2"
rich = ">=13.7,<14"
colorama = "^0.4.6"

[tool.poetry.scripts]
gateway = "gateway.main:main"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
